move schema validation module
