+++ hermes/bin/src/packaging/wasm_module/settings.rs
@@ -2,7 +2,7 @@
 
 use std::io::Read;
 
-use crate::packaging::schema_validation::SchemaValidator;
+use crate::schema_validation::SchemaValidator;
 
 /// Settings schema object.
 #[derive(Debug)]
